Reduce visibility of TextMapCodec.contextFromString to package scope
